% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound.R, R/telegram.R
\name{sound}
\alias{notification}
\alias{sound}
\alias{telegram}
\title{Notifications}
\usage{
sound(sound_file = "extdata/chime.wav")

telegram(user, token = getToken())
}
\arguments{
\item{sound_file}{Location of sound file to play}

\item{user}{Addressee of the notification.}

\item{token}{Character. The bot-specific secret used to access the Telegram API.}
}
\description{
Create a notification. This can be either through sending a Telegram message, an email, by printing a sound, or in its most basic form: printing a notification. A notification is sent through the matching \code{\link{notify}} function.
}
\details{
Below are details of each of the notification types.
}
\section{Sound}{
 Notify by playing a sound
}

\section{Telegram}{

There are two requirements before a notification can be send to a Telegram bot. First, the bot needs to be authorized, which should result in a token. Next, the target user needs to initiate a chat with the bot. If either requirement is not met, the function throws an error.
}
\seealso{
The Telegram bot API at \url{https://core.telegram.org/bots/api#authorizing-your-bot}
}

