% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\alias{notify.default}
\alias{notify.email}
\alias{notify.sound}
\alias{notify.telegram}
\title{Send a notification}
\usage{
notify(notification, ...)

\method{notify}{telegram}(notification, ...)

\method{notify}{email}(notification, ...)

\method{notify}{sound}(notification, ...)

\method{notify}{default}(notification, ...)
}
\arguments{
\item{notification}{A list with at least a message and a start element. Extra elements depend on the type of notification.}
}
\description{
Uses the elements of the \code{notification} object to determine the message to send, and sends it. See \code{\link{notification}} for details of the notification object.

Send a Telegram notification
}
\section{Methods (by class)}{
\itemize{
\item \code{telegram}: Send a Telegram notification

\item \code{email}: Send an email notification

\item \code{sound}: Play a sound notification

\item \code{default}: Print a notification
}}
\examples{
notification <- list("message" = "Done", start = Sys.time())
notify(notification)

\dontrun{
 notification <- telegram("user", token = getToken())
 notification$start = Sys.time()
 notification$message = "Done"
 notify(notification)
}
}

