% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rudone.R
\name{rudone}
\alias{rudone}
\alias{rudone_}
\title{Send a notification after executing some code}
\usage{
rudone(code, notification, results = FALSE)

rudone_(fun, args, notification, results = FALSE)
}
\arguments{
\item{code}{Statements to execute. If there is more than one statement, they should be wrapped in curly brackets.}

\item{notification}{Notification object to send. See \code{\link{notification}} for details.}

\item{results}{Logical. Should the results be send along with the notification?}
}
\description{
There are two versions of this function: \code{rudone} for non-standard evaluation (interactive use), and \code{rudone_} for standard evaluation (programmatic use). Both wrap \code{rudone_exec}.
}
\seealso{
notification
}

