% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rudone.R
\name{rudone_exec}
\alias{rudone_exec}
\title{Execute code and send notification}
\usage{
rudone_exec(expr, notification, results)
}
\arguments{
\item{expr}{Expression.}

\item{notification}{Notification to send after the expression is executed}

\item{results}{Logical. If \code{results == TRUE}, the results of the \emph{final} statement are send along with the notification, if supported by the notification method.}
}
\description{
Executes the expression, and calls \code{\link{notify}} with the notification
}

