% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram.R
\name{token}
\alias{getToken}
\alias{saveToken}
\alias{token}
\title{Retrieve or store a Telegram token}
\usage{
getToken()

saveToken(token)
}
\description{
Looks in \code{extdata/telegram.token} for a Telegram bot token

Saves the token as \code{extdata/telegram.token}
}

